+++ codex/views/browser/browser.py
@@ -51,6 +51,7 @@ class BrowserView(BrowserAnnotationsView):
         {
             AdminFlag.FlagChoices.FOLDER_VIEW.value: "folder_view",
             AdminFlag.FlagChoices.IMPORT_METADATA.value: "import_metadata",
+            AdminFlag.FlagChoices.DYNAMIC_GROUP_COVERS.value: "dynamic_group_covers",
         }
     )