+++ frontend/src/api/v3/cover.js
@@ -1,10 +1,23 @@
 import { getReaderBasePath } from "./common";
 
-export const getCoverSource = (pk, updatedAt) => {
+const COVER_MAP = {
+  p: "publisher",
+  i: "imprint",
+  s: "series",
+  v: "volume",
+  a: "story-arc",
+  f: "folder",
+  c: "missing-cover",
+};
+
+export const getCoverSource = (pk, updatedAt, group) => {
+  let src;
   if (pk) {
     const basePath = getReaderBasePath(pk);
-    return `${basePath}/cover.webp?${updatedAt}`;
+    src = `${basePath}/cover.webp?${updatedAt}`;
   } else {
-    return window.CODEX.MISSING_COVER;
+    const name = COVER_MAP[group];
+    src = window.CODEX.STATIC + "img/" + name + ".svg";
   }
+  return src;
 };